+++ tests/web/WebDriverTest.php
@@ -621,6 +621,14 @@ class WebDriverTest extends TestsForBrowsers
         $this->module->scrollTo('#clickme');
         $this->module->click('Submit');
         $this->module->see('Welcome to test app!');
+    }
 
+    /**
+     * @Issue 2921
+     */
+    public function testSeeInFieldForTextarea()
+    {
+        $this->module->amOnPage('/form/bug2921');
+        $this->module->seeInField('foo', 'bar baz');
     }
 }